+++ ConsoleApplication1/ConsoleApplication1/Transaction.cs
@@ -11,7 +11,7 @@ namespace ConsoleApplication1
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string[] Items { get; set; }
-        public int[] Costs { get; set; }
+        public double[] Costs { get; set; }
         public string Address { get; set; }
         public string Email { get; set; }
         public DateTime Date { get; set; }
@@ -30,7 +30,7 @@ namespace ConsoleApplication1
 
         }
         */
-        public Transaction(string first, string last, string[] items, string address, int[] cost, string email, DateTime date)
+        public Transaction(string first, string last, string[] items, string address, double[] cost, string email, DateTime date)
         {
             FirstName = first;
             LastName = last;
